requirements.txt
flask
sqlalchemy
werkzeug
transformers
torch
sentence-transformers
markdown
faiss-cpu
pillow
PyPDF2
accelerate
